+++ contracts/test/helpers/TestHelpers.sol
@@ -28,7 +28,7 @@ contract TestHelpers {
 
         bytes memory seal = mockVerifier.mockProve(ImageID.RISC0_CALL_GUEST_ID, journalHash).seal;
         Proof memory proof =
-            Proof(journal.length, encodeSeal(seal), 0, [uint256(0), 0, 0, 0, 0, 0, 0, 0, 0, 0], commitment);
+            Proof(journal.length, encodeSeal(seal), 0, [uint16(0), 0, 0, 0, 0, 0, 0, 0, 0, 0], commitment);
         return (proof, journalHash);
     }
